#CAP-15 remove generic and move structs to elgamal.rs,update import.
