+++ src/elgamal.rs
@@ -5,12 +5,35 @@ use rug::ops::Pow;
 use rug::{rand::RandState, Complete, Integer};
 use codec::{Decode, Encode};
 
-use crate::generic::{PublicKey,PrivateKey,Encryption};
 use sp_core::U256;
 use crate::utils::{bigint_u256, u256_bigint};
 
 const STR_RADIX: i32 = 10;
 const SEARCH_LIMIT: u32 = 100;
+
+///Init public key structure for elgamal encryption.
+#[derive(Debug, Clone)]
+pub struct PublicKey {
+    pub p: Integer,
+    pub g: Integer,
+    pub h: Integer,
+    pub bit_length: u32,
+}
+
+///init private key structure for elgamal encryption.
+#[derive(Debug, Clone)]
+pub struct PrivateKey {
+    pub p: Integer,
+    pub g: Integer,
+    pub x: Integer,
+    pub bit_length: u32,
+}
+
+/// A trait to use a RNG and elgamal key to encrypt plaintext to UTF_16LE string.
+pub trait Encryption<I> {
+    fn encrypt(&self,rand: &mut RandState, key: &PublicKey) -> String;
+    fn decrypt(&self,key: &PrivateKey) -> Option<String>;
+}
 
 /// Generate a seed data slice from a key data.
 pub trait Seed {
